Add Postgres collector
Enables streaming replication monitoring by reporting the offset between
current WAL position in master and last received xlog position on slaves
in megabytes.

It is not 100% precise, since connections are established sequentially
to master and followers, but it should be good enough for our purposes.

Example configuration:

    use Collective::Collectors::Postgres, databases: [
      {
        name: "MasterDB",
        url: "<master-db-url>",
        followers: [
          {
            name: "FollowerDB1",
            url: "<follower-db-1-url>"
          },
          {
            name: "FollowerDB2",
            url: "<follower-db-2-url>"
          }
        ]
      }
    ]

Example reported metrics:

    source= measure#postgres.lag_mb.FollowerDB1=0
    source= measure#postgres.lag_mb.FollowerDB2=20